add answer api

+++ personalityTest/src/main/java/com/capstone/personalityTest/repository/TestAttemptRepository.java
@@ -4,7 +4,10 @@ import com.capstone.personalityTest.model.TestAttempt.TestAttempt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface TestAttemptRepository extends JpaRepository<TestAttempt, Integer> {
+public interface TestAttemptRepository extends JpaRepository<TestAttempt, Long> {
+
 
 }
